Iteration 2: fix backend types, add tests, EditEvent, ratings UI, validation
Backend:
- Fix IdentityUser<Guid> type consistency across all entities
- Fix enum vs string comparisons (EventStatus, ParticipantStatus)
- Add DataAnnotation validations on CreateEventDto/UpdateEventDto
- Add rate limiting (auth: 10/min, api: 60/min)
- Make CORS origins configurable via appsettings
- Add EventServiceTests (9 tests) and RatingServiceTests (6 tests)

Frontend:
- Add EditEvent page (/events/:id/edit) with inline validation
- Add RatingForm component (star rating 1-5 + comment)
- Integrate ratings in EventDetail for completed events
- Add client-side validation on CreateEvent (date, title length)
- Add real-time password criteria on Register page
- Setup Vitest + Testing Library (12 tests passing)
- Update README with full stack documentation

+++ backend/PartnR.Api/Services/ProfileService.cs
@@ -55,7 +55,7 @@ public class ProfileService
         Bio = u.Bio,
         AvatarUrl = u.AvatarUrl,
         FavoriteActivities = u.FavoriteActivities,
-        RatingAvg = u.RatingAvg,
+        RatingAvg = (double)u.RatingAvg,
         RatingCount = u.RatingCount,
         CreatedAt = u.CreatedAt
     };
